+++ src/index.js
@@ -19,6 +19,8 @@ import TextEllipsis from './TextEllipsis';
 import TextLabel from './TextLabel';
 import Button from './Button';
 import IconButton from './IconButton';
+import Checkbox from './Checkbox';
+import IconCheckbox from './IconCheckbox';
 import SearchInput from './SearchInput';
 
 export {
@@ -38,5 +40,7 @@ export {
     TextLabel,
     Button,
     IconButton,
+    Checkbox,
+    IconCheckbox,
     SearchInput,
 };